Move some functions
